mv DriverViewWindow to ui/qt/offroad
