+++ selfdrive/ui/qt/home.h
@@ -7,6 +7,7 @@
 #include <QTimer>
 #include <QWidget>
 
+#include "selfdrive/ui/qt/offroad/driverview.h"
 #include "selfdrive/ui/qt/onroad.h"
 #include "selfdrive/ui/qt/sidebar.h"
 #include "selfdrive/ui/qt/widgets/offroad_alerts.h"
@@ -35,29 +36,6 @@ public slots:
   void refresh();
 };
 
-
-class DriverViewWindow : public QOpenGLWidget, protected QOpenGLFunctions {
-  Q_OBJECT
-
-public:
-  using QOpenGLWidget::QOpenGLWidget;
-  explicit DriverViewWindow(QWidget* parent = 0);
-  ~DriverViewWindow();
-
-protected:
-  void paintGL() override;
-  void initializeGL() override;
-
-protected slots:
-  void onTimeout();
-
-private:
-  std::unique_ptr<UIVision> vision;
-  QTimer *timer;
-  QImage face_img;
-  SubMaster sm;
-  bool is_rhd;
-};
 class HomeWindow : public QWidget {
   Q_OBJECT
